requirements.txt
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.7
django-generic-positions==0.2.2
django-heroku==0.3.1
django-hvad==1.8.0
django-libs==2.0.2
django-review==1.10.0
django-user-media==1.2.5
djangorestframework==3.9.0
easy-thumbnails==2.5
peewee==3.7.1
Pillow==5.3.0
psycopg2==2.7.6.1
python-decouple==3.1
pytz==2018.7
simplejson==3.16.0
whitenoise==4.1.1
